separate the node-llamacpp adapter to support the new node bindings
