+++ src/llamacpp/llamacpp-adapter.ts
@@ -5,7 +5,7 @@ export interface LLamaCppPromptOptions {
   onToken?: (chunk: Token[]) => void;
 }
 
-export interface LLamaCppAdaptor {
+export interface LLamaCppAdapter {
   evaluate(query: string, options?: LLamaCppPromptOptions): Promise<string>;
   prompt(text: string, options?: LLamaCppPromptOptions): Promise<string>;
   decode(batch: Token[]): string;
